requirements.txt
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.14
django-allauth==0.42.0
djangorestframework==3.11.0
idna==2.10
ipdb==0.13.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.1
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
